Adding deleteProducts, validation and css to FrontEnd

+++ src/public/js/realTimeProducts.js
@@ -17,14 +17,17 @@ formNewProduct.addEventListener("submit", (e) => {
         thumbnail: document.getElementById("thumbnail").value,
         stock: document.getElementById("stock").value
     }
+    if(!newProduct.title || !newProduct.description || !newProduct.price|| !newProduct.code|| !newProduct.category || !newProduct.thumbnail || !newProduct.stock){
+       return alert('Empty fields, please add all the statements')
+    }
     socket.emit("newProduct", newProduct);
     socket.on('updatedProducts', (data) => {
         try {
             console.log(data);
             document.getElementById("dynamicProducts").innerHTML = data.reduce((acc, item) => {
-                return acc + `<div id=${item.id}>
+                return acc + `<div id=${item.id} style="display: flex; flex-direction: column; justify-content: center; align-items: center; border: solid 3px blue; width: 100%">
                 <h2>${item.title}</h2>
-                <img src="${item.thumbnail}" alt="">
+                <img src="${item.thumbnail}" style="max-width: 100px; max-height: 200px">
                 <h3>Price: ${item.price}</h3>
                 <p>Stock: ${item.stock}</p>
                 <button onclick="deleteProduct(${item.id})" > Delete </button>
@@ -46,9 +49,9 @@ function deleteProduct(id){
         try {
             console.log(data);
             document.getElementById("dynamicProducts").innerHTML = data.reduce((acc, item) => {
-                return acc + `<div id=${item.id}>
+                return acc + `<div id=${item.id} style="display: flex; flex-direction: column; justify-content: center; align-items: center; border: solid 3px blue; width: 100%">
                 <h2>${item.title}</h2>
-                <img src="${item.thumbnail}" alt="">
+                <img src="${item.thumbnail}" style="max-width: 100px; max-height: 200px">
                 <h3>Price: ${item.price}</h3>
                 <p>Stock: ${item.stock}</p>
                 <button onclick="deleteProduct(${item.id})" > Delete </button>
